Extract runBenchmark method
